+++ src/kinetic_sdk/helpers/get_public_key.py
@@ -1,7 +1,7 @@
 # pylint: disable=missing-function-docstring,missing-module-docstring,import-error
 from solana.publickey import PublicKey
 
-from kinetic_sdk.models.public_key_string import PublicKeyString
+from kinetic_sdk.models import PublicKeyString
 
 
 def get_public_key(account: PublicKeyString) -> str:
